Sound Effect
Add an appropriate Sound Effect that triggers when the Floating Platform is Shrinking. You may create a simple sound using any online tool (or download one) for this requirement
Add an appropriate Sound Effect that triggers when the Floating Platform is Resetting. You may create a simple sound using any online tool (or download one) for this requirement